requirements.txt
beautifulsoup4==4.11.1
certifi==2021.10.8
charset-normalizer==2.0.10
colorama==0.4.5
idna==3.3
loguru==0.5.3
netaddr==0.8.0
pyOpenSSL==23.1.1
requests==2.28.1
smbprotocol==1.9.0
urllib3==1.26.8
